Renamed case-summary -> cases-summary
