+++ probation-rules.js
@@ -1,12 +1,12 @@
 const OmWorkload = require('./app/staging/domain/om-workload')
-const CaseSummary = require('./app/staging/domain/case-summary')
+const CasesSummary = require('./app/staging/domain/cases-summary')
 const CaseDetails = require('./app/staging/domain/case-details')
 
 const stagingHelper = require('./test/helpers/staging-helper')
 
 module.exports = {
   OmWorkload: OmWorkload,
-  CaseSummary: CaseSummary,
+  CasesSummary: CasesSummary,
   CaseDetails: CaseDetails,
   stagingTestHelper: stagingHelper
 }
